# src/main/resources/openapi/organization-openapi-v1.0.yml
openapi: 3.1.0
info:
  title: UI-Backend Organization-Employees API
  description: API for interaction between frontend and backend
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
tags:
  - name: organization
    description: Operations about organization
    externalDocs:
      description: Find out more
      url: http://swagger.io
  - name: employee
    description: Operations about employee
    externalDocs:
      description: Find out more about our store
      url: http://swagger.io
paths:
  /api/v1/organization:
    post:
      tags:
        - organization
      summary: Add a new organization
      description: Add a new organization
      operationId: addOrganization
      parameters:
        - name: organizationName
          description: organization name
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      tags:
        - organization
      summary: Get all organizations
      description: For valid response no need to specify parameters
      operationId: getOrganizations
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '404':
          description: organizations not found
    delete:
      tags:
        - organization
      summary: Delete all organizations
      description: For valid response no need to specify parameters
      operationId: deleteOrganizations
      responses:
        '404':
          description: organizations not found

  /api/v1/organization/{organizationId}:
    put:
      tags:
        - organization
      summary: Update an existing organization
      description: Update an existing organization by Id
      operationId: updateOrganizationById
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
        - name: organizationName
          description: organization name
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid ID supplied
        '404':
          description: Organization not found
        '422':
          description: Validation exception
    get:
      tags:
        - organization
      summary: Finds Organization by ID
      description: Get an existing organization by Id
      operationId: getOrganizationById
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid status value
    delete:
      tags:
        - organization
      summary: Delete organization by ID
      description: Delete organization by ID
      operationId: deleteOrganizationById
      parameters:
        - name: organizationId
          in: path
          description: ID of the organization that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Organization not found

  /api/v1/organization/{organizationId}/employee:
    post:
      tags:
        - employee
      summary: Add a new employee to organization
      description: Add a new employee
      operationId: addEmployee
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
        - name: firstName
          description: Employee first name
          schema:
            type: string
          in: path
          required: true
        - name: secondName
          description: Employee second name
          schema:
            type: string
          in: path
          required: true
        - name: jobTitle
          description: Job Title of Employee
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      tags:
        - employee
      summary: Get all employees
      description: For valid response no need to specify parameters
      operationId: getEmployees
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '404':
          description: Employees not found
    delete:
      tags:
        - employee
      summary: Delete all employees
      description: For valid response no need to specify parameters
      operationId: deleteEmployees
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
      responses:
        '404':
          description: Employees not found
  /api/v1/organization/{organizationId}/employee/{employeeId}:
    put:
      tags:
        - employee
      summary: Update employee by organization Id and employee Id
      description: Update a employee
      operationId: updateEmployeeById
      parameters:
        - name: organizationId
          description: organization ID
          schema:
            type: string
          in: path
          required: true
        - name: employeeId
          description: Employee ID
          schema:
            type: string
          in: path
          required: true
        - name: firstName
          description: Employee first name
          schema:
            type: string
          in: path
          required: true
        - name: secondName
          description: Employee second name
          schema:
            type: string
          in: path
          required: true
        - name: jobTitle
          description: Job Title of Employee
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      tags:
        - employee
      summary: Find Employee by ID
      description: Find Employee by ID
      operationId: getEmployeeById
      parameters:
        - name: organizationId
          in: path
          description: ID of the organization
          required: true
          schema:
            type: string
        - name: employeeId
          in: path
          description: ID of the employee
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid Id value
        '404':
          description: Employee not found
    delete:
      tags:
        - employee
      summary: Delete employee by ID
      description: Delete employee by ID
      operationId: deleteEmployeeById
      parameters:
        - name: organizationId
          in: path
          description: ID of the organization that needs to be deleted
          required: true
          schema:
            type: string
        - name: employeeId
          in: path
          description: ID of the employee that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found

components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: string
          nullable: false
          example: b749d000-9a4f-4bb7-8235-3e47779df243
        organizationName:
          type: string
          nullable: false
          description: Organization Name
          example: HHCorp
    Employee:
      type: object
      properties:
        id:
          type: string
          nullable: false
          example: e07bb9d0-d3ef-4c66-9b7d-5b5ee6e47167
        organizationId:
          type: string
          nullable: false
          example: b749d000-9a4f-4bb7-8235-3e47779df243
        firstName:
          type: string
          nullable: false
          example: Test
        secondName:
          type: string
          nullable: false
          example: De La Test
        jobTitle:
          type: string
          nullable: false
          example: Boss